controller/cf.PyControl/requirements.txt
appdirs==1.4.4
blinker==1.8.2
cfclient==2024.11
cflib==0.1.27
click==8.1.7
Flask==3.0.3
freetype-py==2.4.0
hsluv==5.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
libusb-package==1.0.26.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pydot==2.0.0
PyOpenGL==3.1.7
pyparsing==3.1.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pyqtgraph==0.13.7
pyserial==3.5
python-statemachine==2.3.4
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==25.1.2
scipy==1.14.0
setuptools==72.1.0
vispy==0.14.3
waitress==3.0.0
webencodings==0.5.1
Werkzeug==3.0.3
websockets==13.1